moving logic to datalayer
